Move the tooltip message to its own hook
